refactor(data): promote invalidation channel ids + dep keys to @/data/invalidation
The kernel invalidation channel ids (typedBlocks.*, kernel.*) and their
dependency-key builders are a contract plugin queries depend on when they
declare `{kind:'plugin', channel, key}` deps — not an implementation
detail. Move the const + key-builder block out of
`internals/kernelInvalidation.ts` into the public `@/data/invalidation`
module (which already holds the rule/emit machinery types). The kernel's
emit *rules* stay internal and now import the ids from the public module.

Repoint kernelQueries + the find-replace / backlinks / grouped-backlinks
plugins. `emitKernelInvalidations` (kernel-internal) stays where it is.

Part of #139 (promotion-first pass; no lint rule).
